[refactor/postsNetworkBoundResource] refactor: update PostDao and PostEntity to use Long for IDs for consistency

+++ feature/posts/src/main/java/com/alxnophis/jetpack/posts/data/datasource/PostDao.kt
@@ -12,7 +12,7 @@ internal interface PostDao {
     suspend fun getAllPosts(): List<PostEntity>
 
     @Query("SELECT * FROM posts WHERE id = :postId")
-    suspend fun getPostById(postId: Int): PostEntity?
+    suspend fun getPostById(postId: Long): PostEntity?
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     suspend fun insertPosts(posts: List<PostEntity>)
