+++ feature/posts/src/main/java/com/alxnophis/jetpack/posts/data/model/PostEntity.kt
@@ -7,7 +7,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "posts")
 internal data class PostEntity(
     @PrimaryKey @ColumnInfo(name = "id") val id: Long,
-    @ColumnInfo(name = "user_id") val userId: Int,
+    @ColumnInfo(name = "user_id") val userId: Long,
     @ColumnInfo(name = "title") val title: String,
     @ColumnInfo(name = "body") val body: String,
 )
